Added: PropertyMap which is the list of properties added as a Row attribute. This means that Row, when serialized, will have a new property {"p": xxxx} Added: New code (tests) that validates the correctness of the JSON returned.
Bug Fix: Removed a trailing comma in the JSON string when a Cell  (p tag) is serialized. This was causing an invalid JSON string
Bug Fix: Changed the way of how the Date object get's serialized in the JSON string. Now the object is returned quoted. "new Date..."
Bug Fix: Default column Id now starts with an index of 0 rather than 1.